refactor: mode default routes to seperate file
